% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_tokenizer.R
\name{model_to_tokenizer}
\alias{model_to_tokenizer}
\title{Gets the name of the tokenizer used by a model}
\usage{
model_to_tokenizer(model)
}
\arguments{
\item{model}{the model to use, e.g., \verb{gpt-4o}}
}
\value{
the tokenizer used by the model
}
\description{
Gets the name of the tokenizer used by a model
}
\examples{
model_to_tokenizer("gpt-4o")
model_to_tokenizer("gpt-4-1106-preview")
model_to_tokenizer("text-davinci-002")
model_to_tokenizer("text-embedding-ada-002")
model_to_tokenizer("text-embedding-3-small")
}
