% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{get_tokens}
\alias{get_tokens}
\title{Converts text to tokens}
\usage{
get_tokens(text, model, verbose = FALSE)
}
\arguments{
\item{text}{a character string to encode to tokens, can be a vector}

\item{model}{a model to use for tokenization, either a model name, eg \verb{gpt-4o}
or a tokenizer, eg \code{o200k_base}.
See also \href{https://github.com/zurawiki/tiktoken-rs/blob/main/tiktoken-rs/src/tokenizer.rs}{available tokenizers}.}

\item{verbose}{if \code{TRUE} print additional information, default \code{FALSE}}
}
\value{
a vector of tokens for the given text as integer
}
\description{
Converts text to tokens
}
\examples{
get_tokens("Hello World", "gpt-4o")
get_tokens("Hello World", "o200k_base")
get_tokens("Hello World", "clearly wrong")
}
\seealso{
\url{model_to_tokenizer}
}
